+++ src/main/java/ec/com/expensys/persistence/repository/ExpRolePersonRepository.java
@@ -3,5 +3,5 @@ package ec.com.expensys.persistence.repository;
 import ec.com.expensys.persistence.entity.ExpRolePerson;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ExpRolePersonRepository extends CrudRepository<ExpRolePerson, Integer> {
+public interface ExpRolePersonRepository extends CrudRepository<ExpRolePerson, Long> {
 }
