+++ src/main/java/ec/com/expensys/persistence/repository/ExpRoleRepository.java
@@ -3,6 +3,6 @@ package ec.com.expensys.persistence.repository;
 import ec.com.expensys.persistence.entity.ExpRole;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ExpRoleRepository extends CrudRepository<ExpRole, Integer> {
+public interface ExpRoleRepository extends CrudRepository<ExpRole, Long> {
 
 }
