+++ src/main/java/ec/com/expensys/persistence/repository/ExpTransactionRepository.java
@@ -3,5 +3,5 @@ package ec.com.expensys.persistence.repository;
 import ec.com.expensys.persistence.entity.ExpTransaction;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ExpTransactionRepository extends CrudRepository<ExpTransaction, Integer> {
+public interface ExpTransactionRepository extends CrudRepository<ExpTransaction, Long> {
 }
